requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==5.0.7
django-ckeditor==6.7.1
django-crispy-forms==2.2
django-js-asset==2.2.0
idna==3.7
pillow==10.4.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
